apis/requirements.txt
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
django-cors-headers==3.7.0
python-ldap==3.3.1